// FootballMathesCollectorClient/src/app/models/match.model.ts
export class Match {
    id?: any;
    date?: string;
    result?: string;
    homeClub?:string;
    awayClub?: string;
    awayGoalkeeper?: string;
    awayLeftBack?: string;
    awayLeftCenterBack?: string;
    awayRightCenterBack?: string;
    awayRightBack?: string;
    awayLeftMidfielder?: string;
    awayLeftCenterMidfielder?: string;
    awayRightCenterMidfielder?: string;
    awayRightMidfielder?: string;
    awayLeftStriker?: string;
    awayRightStriker?: string;
    homeGoalkeeper?: string;
    homeLeftBack?: string;
    homeLeftCenterBack?: string;
    homeRightCenterBack?: string;
    homeRightBack?: string;
    homeLeftMidfielder?: string;
    homeLeftCenterMidfielder?: string;
    homeRightCenterMidfielder?: string;
    homeRightMidfielder?: string;
    homeLeftStriker?: string;
    homeRightStriker?: string;
}